Level-0: Rename Duke to Zbot, implement greeting and exit
- Renamed class from Duke to Zbot
- Renamed file from Duke.java to Zbot.java
- Added greeting message "Hello! I'm Zbot"
- Added "What can I do for you?" prompt
- Added exit message "Bye. Hope to see you again soon!"
- Basic user input handling with Scanner
